uefi: Move PoolString to enable additional use
